requirements.txt
supabase
pyqrcode
pypng
python-dotenv
Flask
